More standardized module structure and clt
